Refactor: adjusting location of PERSON validations

+++ src/main/java/com/electricitybill/service/impl/PessoaServiceImpl.java
@@ -3,10 +3,12 @@ package com.electricitybill.service.impl;
 import com.electricitybill.entity.PessoaEntity;
 import com.electricitybill.repository.PessoaRepository;
 import com.electricitybill.service.PessoaService;
+import jakarta.persistence.NoResultException;
 import lombok.AllArgsConstructor;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Objects;
 
 @AllArgsConstructor
 public class PessoaServiceImpl implements PessoaService {
@@ -15,11 +17,17 @@ public class PessoaServiceImpl implements PessoaService {
 
     @Override
     public List<PessoaEntity> findAll() {
+        if (pessoaRepository.findAll().isEmpty()) {
+            throw new NoResultException("Search not found!");
+        }
         return pessoaRepository.findAll();
     }
 
     @Override
     public PessoaEntity findById(int id) {
+        if (Objects.isNull(pessoaRepository.findById(id))) {
+            throw new NoResultException("Individual search not found!");
+        }
         return pessoaRepository.findById(id);
     }
 
@@ -30,11 +38,17 @@ public class PessoaServiceImpl implements PessoaService {
 
     @Override
     public boolean update(int id, PessoaEntity entity) throws SQLException {
+        if (Objects.isNull(pessoaRepository.findById(id))) {
+            throw new NoResultException("Person not found!");
+        }
         return pessoaRepository.update(id, entity);
     }
 
     @Override
     public boolean delete(int id) throws SQLException {
+        if (Objects.isNull(pessoaRepository.findById(id))) {
+            throw new NoResultException("Person not found!");
+        }
         return pessoaRepository.delete(id);
     }
 }
